Reactivate layerFilterSubsetString and deactivate layerFilterExpression
Because of `QgsAccessControl` stored filter expressions with resolveFilterFeatures used in WMS
and does not clear filter expressions cached at the end of request.

Because of `QgsFeatureRequest` could not be `FilterExpression` and `FilterFids`, `QgsAccessControl`
filter expressions should not be apply to QgsFeatureRequest when it's type is `FilterFids`.

Lizmap cannot use filter expressions and has to use filter subsetstring.