[Bugfix] GetLegendGraphic JSON with SHOWFEATURECOUNT
When the request contains SHOWFEATURECOUNT parameter, the symbol title is not equal to the symbol label.
To fix it, the symbol label is extracted from the symbol title when the request contains SHOWFEATURECOUNT.

With QGIS 3.22, it seems that countSymbolFeatures is not used for JSON output.
To fix it, countSymbolFeatures is used.